[Refactor] 온보딩 및 Auth 관련 코드 리펙토링
