+++ src/main/java/com/grabpt/service/CategoryService/CategoryQueryService.java
@@ -1,12 +1,11 @@
 package com.grabpt.service.CategoryService;
 
-import com.grabpt.domain.entity.Category;
-import com.grabpt.domain.entity.Requestions;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
 import java.util.List;
+
+import com.grabpt.domain.entity.Category;
 
 public interface CategoryQueryService {
 	List<Category> getCategories();
+
+	Category findById(long id);
 }
