+++ src/main/java/com/grabpt/service/CategoryService/CategoryQueryServiceImpl.java
@@ -1,23 +1,29 @@
 package com.grabpt.service.CategoryService;
 
-import com.grabpt.domain.entity.Category;
-import com.grabpt.domain.entity.Requestions;
-import com.grabpt.repository.CategoryRepository.CategoryRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.grabpt.apiPayload.code.status.ErrorStatus;
+import com.grabpt.apiPayload.exception.handler.CategoryHandler;
+import com.grabpt.domain.entity.Category;
+import com.grabpt.repository.CategoryRepository.CategoryRepository;
+
+import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class CategoryQueryServiceImpl implements CategoryQueryService{
+public class CategoryQueryServiceImpl implements CategoryQueryService {
 
 	private final CategoryRepository categoryRepository;
 
 	@Override
 	public List<Category> getCategories() {
 		return categoryRepository.findAll();
+	}
+
+	@Override
+	public Category findById(long id) {
+		return categoryRepository.findById(id).orElseThrow(() -> new CategoryHandler(ErrorStatus.CATEGORY_NOT_FOUND));
 	}
 }
